refactor : follow 구조 변경
